More fixes after working with Greg
