+++ Sources/SpotHeroAPI/Search/Monthly/Models/MonthlyRate.swift
@@ -76,21 +76,6 @@ public struct MonthlyRate: Codable {
     public let isOversized: Bool
 }
 
-public extension MonthlyRate {
-    struct ParkingPass: Codable {
-        private enum CodingKeys: String, CodingKey {
-            case displayName = "display_name"
-            case type
-        }
-        
-        /// The display name for the parking pass to be shown to users.
-        public let displayName: String
-        
-        /// Defines the supported parking pass types
-        public let type: String
-    }
-}
-
 // MARK: - Enums
 
 public extension MonthlyRate {
