+++ Tests/SpotHeroAPITests/Core/APITestCase.swift
@@ -7,7 +7,7 @@ import XCTest
 class APITestCase: XCTestCase {
     enum ServiceURL: String {
         case monolith = "https://mobile.staging.spothero.com"
-        case craig = "https://api.staging.spothero.com/mobile"
+        case craig = "https://api.staging.spothero.com"
     }
     
     static var timeout: TimeInterval = 15
